Consolidate samplers
