--- opencensus/trace/samplers/always_on.py
@@ -1,23 +1,0 @@
-# Copyright 2017, OpenCensus Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from opencensus.trace.samplers.base import Sampler
-
-
-class AlwaysOnSampler(Sampler):
-    """Trace every request."""
-
-    def should_sample(self, trace_id):
-        """Always return True because we want to sample all requests."""
-        return True
