+++ tests/unit/trace/samplers/test_probability.py
@@ -14,50 +14,50 @@
 
 import unittest
 
-from opencensus.trace.samplers import probability
+from opencensus.trace import samplers
 
 
 class TestProbabilitySampler(unittest.TestCase):
     def test_constructor_invalid(self):
         with self.assertRaises(ValueError):
-            probability.ProbabilitySampler(rate=2)
+            samplers.ProbabilitySampler(rate=2)
 
     def test_constructor_valid(self):
         rate = 0.8
-        sampler = probability.ProbabilitySampler(rate=rate)
+        sampler = samplers.ProbabilitySampler(rate=rate)
 
         self.assertEqual(sampler.rate, rate)
 
     def test_constructor_default(self):
         rate = 0.5
-        sampler = probability.ProbabilitySampler()
+        sampler = samplers.ProbabilitySampler()
 
         self.assertEqual(sampler.rate, rate)
 
     def test_should_sample_smaller(self):
         trace_id = 'f8739df974a4481f98748cd92b27177d'
-        sampler = probability.ProbabilitySampler(rate=1)
+        sampler = samplers.ProbabilitySampler(rate=1)
         should_sample = sampler.should_sample(trace_id=trace_id)
 
         self.assertTrue(should_sample)
 
     def test_should_sample_greater(self):
         trace_id = 'f8739df974a4481f98748cd92b27177d'
-        sampler = probability.ProbabilitySampler(rate=0)
+        sampler = samplers.ProbabilitySampler(rate=0)
         should_sample = sampler.should_sample(trace_id=trace_id)
 
         self.assertFalse(should_sample)
 
     def test_should_sample_trace_id_sampled(self):
         trace_id = '00000000000000000000000000000000'
-        sampler = probability.ProbabilitySampler(rate=0.5)
+        sampler = samplers.ProbabilitySampler(rate=0.5)
         should_sample = sampler.should_sample(trace_id=trace_id)
 
         self.assertTrue(should_sample)
 
     def test_should_sample_trace_id_not_sampled(self):
         trace_id = 'ffffffffffffffffffffffffffffffff'
-        sampler = probability.ProbabilitySampler(rate=0.5)
+        sampler = samplers.ProbabilitySampler(rate=0.5)
         should_sample = sampler.should_sample(trace_id=trace_id)
 
         self.assertFalse(should_sample)
